// src/types/currency.ts
export interface Currency {
    id: string
    name: string
    symbol: string
}

export type CurrencyCode =
    | "USD"
    | "EUR"
    | "GBP"
    | "CNY"
    | "JPY"
    | "CAD"
    | "AUD"
    | "HKD"
    | "ISK"
    | "PHP"
    | "DKK"
    | "HUF"
    | "CZK"
    | "RON"
    | "SEK"
    | "IDR"
    | "INR"
    | "BRL"
    | "RUB"
    | "HRK"
    | "THB"
    | "CHF"
    | "MYR"
    | "BGN"
    | "TRY"
    | "NOK"
    | "NZD"
    | "ZAR"
    | "MXN"
    | "SGD"
    | "ILS"
    | "KRW"
    | "PLN"
    | "AED"
    | "ARS"
    | "CLP"
    | "COP"
    | "CRC"
    | "KWD"
    | "KZT"
    | "PEN"
    | "QAR"
    | "SAR"
    | "TWD"
    | "UAH"
    | "UYU"
    | "VND"
    | "GEL"
    | "BTC"
    | "ETH"
    | "FET"